requirements.txt
Pillow==10.0.0
plotly==5.24.1
streamlit==1.41.1
